Minor update to sendStruct to handle a potential error case
Previously, if you used a matrix or a column vector, sendStruct would
get an error on the m=double()  line. Now, we make sure that everything
is converted to a row vector prior to that line. We really shouldn't
put send matrices this way because the dimensionality is lost, but at
least the numbers get across correctly. But for column vectors, this is
fine and avoids the error.