changes to runes to fix allCodes writing and sendStruct issues
This version overwrites the changes in a recent commit related to doing
some error correction on the msg passing. It seems we may be switching
to serial anyway. For now, what this version does is:
1 - add the calibration info to params so it gets saved with the data
2 - updates some try/catch statements to throw the proper error
3 - fixes the trialTic and thisTrialCodes initialization so it is
properly setup for sendStruct, which now puts the codes into trialCodes
in addition to sending on the digital port
4 - makes sure the codes in allCodes are set so that code 1 (trial
start) is at time zero, and all other codes in that trial are timed
relative to it.
5 - maybe a few other minor things related to the above